Use add button instead of remove in search results

+++ app/helpers/friendships_helper.rb
@@ -1,5 +1,5 @@
 module FriendshipsHelper
-  def friends_list_item(friend)
+  def friends_list_item(friend, search = false)
     content_tag(:li, class: 'list-group-item') do
       content_tag(:div, class: 'row') do
         content_tag(:div, class: 'col-md-10') do
@@ -15,7 +15,12 @@ module FriendshipsHelper
           end
         end +
         content_tag(:div, class: 'col-md-2') do
-          link_to 'Remove', friendship_path(friend), method: :delete, remote: true, class: 'btn btn-danger delete-friendship'
+          if search
+            link_to 'Add', requests_path(requester_id: current_user.id, requested_id: friend.id),
+              method: :post, remote: true, class: 'btn btn-success add-friendship'
+          else
+            link_to 'Remove', friendship_path(friend), method: :delete, remote: true, class: 'btn btn-danger delete-friendship'
+          end
         end
       end
     end
